Removes Checked Headers, Instead Includes Headers from Spec Repository (#44)
This adds a `CHECKEDC_SPEC_DIR` variable to point to where we've checked out Microsoft/checkedc

It will then add the `include` dir in that repository to the include path, so that we can use the checked c headers there. We also remove the `include` dir and its contents from this repository to avoid confusion. We didn't change the Makefiles before, so I'm not changing them now. The previous change to the CMake configuration by us was this: 5c13ec2d7611605e17f35c863a11469784fb95fb

This removes the copy of the headers we have in this repository, because syncing changes is bound to mean we lose changes between both repositories.

This is the diff to my benchmarking scripts that was needed: https://github.com/lenary/checkedc-travis-builder/commit/b5432c68131004d14d7cf525d8d9eb7edcc59f9a

Tested:
- On My Mac
- In the Benchmarking system, using that change to the scripts above.