Benchmark Conversion of Ptrdist/yacr2 (#65)
The overall quality of code in this benchmark was infuriating. Lots of globals, lots passed as parameters sometimes but not always. Lots of parameter bounds contain lvalue reads of globals. If we want to avoid this we're going to be passing a lot more parameters everywhere. I had to introduce an intermediate struct in one place.

I don't really have an idea how much slower this is than its pre-converted form. I'll get an idea when I re-run the benchmarks.

Closes #48