Remove workaround from benchmark. (#85)
The compiler isn't properly inferring bounds of count(0) for return interface types of the form `itype(_Nt_array_ptr)`.   We had to put a workaround for the use of `strtok`, which returns an `_Nt_array_ptr`.   With https://github.com/Microsoft/checkedc-clang/pull/461, we can put both  an interface type and a bounds on a return type, and declare the proper bounds for `strtok'.   We can remove the workaround.