Fix incorrect cast in Ptrdist\ks (#94)
There was a cast from a checked array to an nt_array_ptr in a checked scope in Ptrdist\ks. This is now disallowed by the compiler. Change the checked array to a null-terminated checked array, to eliminate the need for the cast. Also remove a cast involving strtok that worked around a compiler bug that has been fixed.