Fix Olden and Ptrdist benchmarks (#98)
This PR updates test files so that they compile without errors, to account for new compiler errors introduced by checkedc-clang/853.

If a variable v is used in the bounds of a variable p, and v is modified without an original value, then the observed bounds of v will be unknown. This is a compiler error as of checkedc-clang/853 that occurred in two llvm test suite files.