Fix failing tests after Clang 11 upgrade (#104)
* [test-suite] Add -fcommon

D75056 is going change defaults from -fcommon to -fno-common. This had
to be reverted because of these failing  test cases:

  FAIL: MultiSource/Applications/JM/ldecod/ldecod.compile_time (1 of 2630)
  FAIL: MultiSource/Applications/JM/lencod/lencod.compile_time (2 of 2630)
  FAIL: MultiSource/Benchmarks/ASC_Sequoia/IRSmk/IRSmk.compile_time (3 of 2630)
  FAIL: MultiSource/Benchmarks/DOE-ProxyApps-C/miniAMR/miniAMR.compile_time (4 of 2630)
  FAIL: MultiSource/Benchmarks/Olden/bh/bh.compile_time (5 of 2630)
  FAIL: MultiSource/Benchmarks/Prolangs-C/TimberWolfMC/timberwolfmc.compile_time (6 of 2630)
  FAIL: MultiSource/Benchmarks/Prolangs-C/compiler/compiler.compile_time (7 of 2630)
  FAIL: MultiSource/Benchmarks/Prolangs-C/loader/loader.compile_time (8 of 2630)
  FAIL: MultiSource/Benchmarks/Prolangs-C/simulator/simulator.compile_time (9 of 2630)

which all failed compiling with "multiple definition of .. " link errors as
a result of this change. So, this adds -fcommon to the CMake files that need this,
and is a prep step for recommiting D75056.

Differential revision: https://reviews.llvm.org/D75557

* [test-suite] switch Clang driver to default to -fno-common

I've tried recommiting the Clang patch to switch to -fno-common, but
again some of the build bots turned red running the test-suite. After my
first commit, I added -fcommon to CMakeLists.txt for some benchmarks
that need this, see my previous test-suite commit. After the recommit,
and the same tests failing again on some bots, I realised that some bots
use Make, so this adds -fcommon to also their Makefiles.

Co-authored-by: Sjoerd Meijer <sjoerd.meijer@arm.com>